v0.2.0: auto port selection

+++ example/multiproc_multi.py
@@ -1,13 +1,14 @@
 import logging
+import multiprocessing
 import time
 from multiprocessing import Process
 
-from example.multi_judge import PongJudgeEnv
 from example.multi_agent import PongAgentEnv
+from example.multi_judge import PongJudgeEnv
 
 
-def run_agent(uid):
-    env = PongAgentEnv(uid=uid)
+def run_agent(uid, port):
+    env = PongAgentEnv(uid=uid, port=port)
     for i_episode in range(2):
         env.reset()
         for t in range(10000):
@@ -21,18 +22,30 @@ def run_agent(uid):
     env.close()
 
 
-def main():
+def run_judge(return_queue):
     judge_env = PongJudgeEnv()
-    judge_proc = Process(target=judge_env.start, args=())
+    return_queue.put(judge_env.bind())
+    judge_env.start()
+
+
+def main():
+    manager = multiprocessing.Manager()
+    return_queue = manager.Queue()
+    judge_proc = Process(target=run_judge, args=(return_queue,))
     judge_proc.start()
-    agent_proc_0 = Process(target=run_agent, args=(0,))
-    agent_proc_1 = Process(target=run_agent, args=(1,))
-    agent_proc_0.start()
-    agent_proc_1.start()
-    agent_proc_0.join()
-    agent_proc_1.join()
-    judge_proc.kill()
-    judge_env.close()
+    for _ in range(10):  # wait for up to 10 seconds
+        if not return_queue.empty():
+            port = return_queue.get()
+            agent_proc_0 = Process(target=run_agent, args=(0, port))
+            agent_proc_1 = Process(target=run_agent, args=(1, port))
+            agent_proc_0.start()
+            agent_proc_1.start()
+            agent_proc_0.join()
+            agent_proc_1.join()
+            judge_proc.kill()
+            break
+        else:
+            time.sleep(1)
 
 
 if __name__ == "__main__":
